initial 3.9Nr patches

+++ lib/Xt/Convert.c
@@ -32,7 +32,7 @@ OTHER TORTIOUS ACTION, ARISING OUT OF OR IN CONNECTION  WITH
 THE USE OR PERFORMANCE OF THIS SOFTWARE.
 
 ******************************************************************/
-/* $XFree86: xc/lib/Xt/Convert.c,v 3.2 1998/01/11 03:48:14 dawes Exp $ */
+/* $XFree86: xc/lib/Xt/Convert.c,v 3.3 1998/10/03 09:06:50 dawes Exp $ */
 
 /*
 
@@ -612,7 +612,8 @@ static void ComputeArgs(widget, convert_args, num_args, args)
 	    args[i].addr =
 		(XPointer)((int)widget + (int)convert_args[i].address_id);
 #else
-	    args[i].addr = (XPointer)((char *)widget + (int)convert_args[i].address_id);
+	    args[i].addr =
+		(XPointer)((char *)widget + (long)convert_args[i].address_id);
 #endif
 	    break;
 
@@ -629,7 +630,7 @@ static void ComputeArgs(widget, convert_args, num_args, args)
 		(XPointer)((int)ancestor + (int)convert_args[i].address_id);
 #else
 	    args[i].addr =
-		(XPointer)((char *)ancestor + (int)convert_args[i].address_id);
+		(XPointer)((char *)ancestor + (long)convert_args[i].address_id);
 #endif
 	    break;
 
@@ -646,14 +647,14 @@ static void ComputeArgs(widget, convert_args, num_args, args)
 	    /* Convert in place for next usage */
 	    convert_args[i].address_mode = XtResourceQuark;
 	    convert_args[i].address_id =
-	       (XtPointer)XrmStringToQuark((String)convert_args[i].address_id);
+	       (XtPointer)(long)XrmStringToQuark((String)convert_args[i].address_id);
 	    /* Fall through */
 
 	case XtResourceQuark:
 	    if (! ResourceQuarkToOffset(widget->core.widget_class,
-		    (XrmQuark) convert_args[i].address_id, &offset)) {
+		    (XrmQuark)(long) convert_args[i].address_id, &offset)) {
 		params[0]=
-                  XrmQuarkToString((XrmQuark) convert_args[i].address_id);
+                  XrmQuarkToString((XrmQuark)(long) convert_args[i].address_id);
                XtAppWarningMsg(XtWidgetToApplicationContext(widget),
 		    "invalidResourceName","computeArgs",XtCXtToolkitError,
 		    "Cannot find resource name %s as argument to conversion",
@@ -691,7 +692,7 @@ void XtDirectConvert(converter, args, num_args, from, to)
 
     LOCK_PROCESS;
     /* Try to find cache entry for conversion */
-    hash = ((int)(converter) >> 2) + from->size + *((char *) from->addr);
+    hash = ((long) converter >> 2) + from->size + *((char *) from->addr);
     if (from->size > 1) hash += ((char *) from->addr)[1];
     
     for (p = cacheHashTable[hash & CACHEHASHMASK]; p; p = p->next) {
@@ -789,7 +790,7 @@ CallConverter(dpy, converter,
 
     LOCK_PROCESS;
     /* Try to find cache entry for conversion */
-    hash = ((int)(converter) >> 2) + from->size + *((char *) from->addr);
+    hash = ((long)(converter) >> 2) + from->size + *((char *) from->addr);
     if (from->size > 1) hash += ((char *) from->addr)[1];
     
     if (cP->cache_type != XtCacheNone) {
