+++ programs/Xserver/mi/mibstore.c
@@ -38,7 +38,7 @@ implied warranty.
 
 ******************************************************************/
 
-/* $XFree86: xc/programs/Xserver/mi/mibstore.c,v 1.2 1998/04/05 16:42:28 robin Exp $ */
+/* $XFree86: xc/programs/Xserver/mi/mibstore.c,v 1.3 1998/10/04 09:39:24 dawes Exp $ */
 
 #define NEED_EVENTS
 #include "X.h"
@@ -2302,8 +2302,8 @@ miBSClearBackingStore(pWin, x, y, w, h, generateExposures)
 		    gcvalues[1] = (pointer) background.pixmap;
 		    gcmask = GCFillStyle|GCTile;
 		}
-		gcvalues[2] = (pointer)(ts_x_origin - pBackingStore->x);
-		gcvalues[3] = (pointer)(ts_y_origin - pBackingStore->y);
+		gcvalues[2] = (pointer)(long)(ts_x_origin - pBackingStore->x);
+		gcvalues[3] = (pointer)(long)(ts_y_origin - pBackingStore->y);
 		gcmask |= GCTileStipXOrigin|GCTileStipYOrigin;
 		DoChangeGC(pGC, gcmask, (XID *)gcvalues, TRUE);
 		ValidateGC((DrawablePtr)pBackingStore->pBackingPixmap, pGC);
@@ -2446,12 +2446,12 @@ miBSFillVirtualBits (pDrawable, pGC, pRgn, x, y, state, pixunion, planeMask)
 	}
 	if (pGC->patOrg.x != x)
 	{
-	    gcval[i++] = (pointer)x;
+	    gcval[i++] = (pointer)(long)x;
 	    gcmask |= GCTileStipXOrigin;
 	}
 	if (pGC->patOrg.y != y)
 	{
-	    gcval[i++] = (pointer)y;
+	    gcval[i++] = (pointer)(long)y;
 	    gcmask |= GCTileStipYOrigin;
 	}
     }
